pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "arc-agi-benchmarking"
version = "0.1.0"
license = {file = "LICENSE.md"}
classifiers = [
    "License :: OSI Approved :: BSD License"
]

dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "click==8.1.8",
    "distro==1.9.0",
    "filelock==3.16.1",
    "fsspec==2024.10.0",
    "h11==0.14.0",
    "httpcore==1.0.6",
    "httplib2==0.22.0",
    "httpx==0.27.2",
    "huggingface-hub==0.29.1",
    "idna==3.10",
    "iniconfig==2.0.0",
    "jiter==0.6.1",
    "packaging==24.1",
    "pluggy==1.5.0",
    "pydantic==2.9.2",
    "pydantic_core==2.23.4",
    "pyparsing==3.2.1",
    "pytest==8.3.5",
    "python-dotenv==1.0.1",
    "PyYAML==6.0.2",
    "sniffio==1.3.1",
    "tqdm==4.66.5",
    "typing_extensions==4.12.2",
    "urllib3==2.2.3",
]

description = "Code for testing model on ARC-AGI-1 and ARC-AGI-2"
readme = "README.md"


# Optional dependencies
[project.optional-dependencies]

providers = [
    "anthropic==0.49.0",
    "cachetools==5.5.2",
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.24.1",
    "google-api-python-client==2.162.0",
    "google-auth==2.38.0",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.8.4",
    "googleapis-common-protos==1.68.0",
    "grpcio==1.60.1",
    "grpcio-status==1.60.1",
    "openai==1.75.0",
    "proto-plus==1.26.0",
    "protobuf==5.29.3",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.1",
    "requests==2.32.3",
    "rsa==4.9",
    "uritemplate==4.1.1"
]


[project.urls]
Home = "https://github.com/arcprize/model_baseline"